Only mark the first sample of a post-roll recovery as a member of roll-group.
This change doesn't follow the description in H.264/AVC Textbook (Sakae Okubo  (ed.), et al.),
but is more closer to the definition in the specification, I think.
If this implementation is wrong, please tell me or us.

+++ lsmash.h
@@ -1113,7 +1113,7 @@ typedef struct
 
 typedef struct
 {
-    uint16_t distance;      /* pre-roll distance for representing audio decoder delay derived from composition
+    uint32_t distance;      /* pre-roll distance for representing audio decoder delay derived from composition
                              * For example, typical AAC encoding uses a transform over consecutive sets of 2048 audio samples,
                              * applied every 1024 audio samples (MDCTs are overlapped).
                              * For correct audio to be decoded, both transforms for any period of 1024 audio samples are needed.
